channels_tut/requirements.txt
channels==0.17.2
asgi_redis==0.14.1
dj_database_url==0.4.1
djangorestframework==3.4.6
redis==2.10.5
psycopg2==2.6.2
gunicorn==19.6.0